y2-t3_keep-playing/keep-playing-master/keep-playing-master/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.24.17
botocore==1.27.17
dj-database-url==0.5.0
Django==4.0.5
django-environ==0.8.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
docx2pdf==0.1.8
extensions==0.4
gunicorn==20.1.0
jmespath==1.0.1
lxml==4.9.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
pytz==2021.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
whitenoise==6.2.0
